+++ internal/auth/subtype.go
@@ -3,7 +3,9 @@ package auth
 import (
 	"strings"
 
+	"github.com/hashicorp/boundary/internal/auth/oidc"
 	"github.com/hashicorp/boundary/internal/auth/password"
+	"github.com/hashicorp/boundary/internal/db/timestamp"
 )
 
 type SubType int
@@ -11,15 +13,50 @@ type SubType int
 const (
 	UnknownSubtype SubType = iota
 	PasswordSubtype
+	OidcSubtype
 )
 
 func (t SubType) String() string {
 	switch t {
 	case PasswordSubtype:
 		return "password"
+	case OidcSubtype:
+		return "oidc"
 	}
 	return "unknown"
 }
+
+// AuthMethod contains the common methods across all the different types of auth methods.
+type AuthMethod interface {
+	GetPublicId() string
+	GetCreateTime() *timestamp.Timestamp
+	GetUpdateTime() *timestamp.Timestamp
+	GetName() string
+	GetDescription() string
+	GetScopeId() string
+	GetVersion() uint32
+	GetIsPrimaryAuthMethod() bool
+}
+
+var (
+	_ AuthMethod = (*oidc.AuthMethod)(nil)
+	_ AuthMethod = (*password.AuthMethod)(nil)
+)
+
+type Account interface {
+	GetPublicId() string
+	GetCreateTime() *timestamp.Timestamp
+	GetUpdateTime() *timestamp.Timestamp
+	GetName() string
+	GetDescription() string
+	GetAuthMethodId() string
+	GetVersion() uint32
+}
+
+var (
+	_ Account = (*oidc.Account)(nil)
+	_ Account = (*password.Account)(nil)
+)
 
 // SubtypeFromType converts a string to a SubType.
 // returns UnknownSubtype if no SubType with that name is found.
@@ -27,6 +64,8 @@ func SubtypeFromType(t string) SubType {
 	switch {
 	case strings.EqualFold(strings.TrimSpace(t), PasswordSubtype.String()):
 		return PasswordSubtype
+	case strings.EqualFold(strings.TrimSpace(t), OidcSubtype.String()):
+		return OidcSubtype
 	}
 	return UnknownSubtype
 }
@@ -39,6 +78,9 @@ func SubtypeFromId(id string) SubType {
 	case strings.HasPrefix(strings.TrimSpace(id), password.AuthMethodPrefix),
 		strings.HasPrefix(strings.TrimSpace(id), password.AccountPrefix):
 		return PasswordSubtype
+	case strings.HasPrefix(strings.TrimSpace(id), oidc.AuthMethodPrefix),
+		strings.HasPrefix(strings.TrimSpace(id), oidc.AccountPrefix):
+		return OidcSubtype
 	}
 	return UnknownSubtype
 }