+++ internal/cmd/commands/authtokenscmd/authtokens.gen.go
@@ -165,7 +165,7 @@ func (c *Command) Run(args []string) int {
 
 	var version uint32
 
-	if ok := extraFlagsHandlingFunc(c, &opts); !ok {
+	if ok := extraFlagsHandlingFunc(c, f, &opts); !ok {
 		return base.CommandUserError
 	}
 
@@ -279,7 +279,7 @@ var (
 	extraActionsFlagsMapFunc = func() map[string][]string { return nil }
 	extraSynopsisFunc        = func(*Command) string { return "" }
 	extraFlagsFunc           = func(*Command, *base.FlagSets, *base.FlagSet) {}
-	extraFlagsHandlingFunc   = func(*Command, *[]authtokens.Option) bool { return true }
+	extraFlagsHandlingFunc   = func(*Command, *base.FlagSets, *[]authtokens.Option) bool { return true }
 	executeExtraActions      = func(_ *Command, inResult api.GenericResult, inErr error, _ *authtokens.Client, _ uint32, _ []authtokens.Option) (api.GenericResult, error) {
 		return inResult, inErr
 	}