+++ internal/db/schema/manager.go
@@ -6,6 +6,7 @@ import (
 	"database/sql"
 	"fmt"
 	"io"
+	"time"
 
 	"github.com/hashicorp/boundary/internal/db/schema/postgres"
 	"github.com/hashicorp/boundary/internal/errors"
@@ -38,16 +39,25 @@ type driver interface {
 // the underlying boundary database schema.
 // Manager is not thread safe.
 type Manager struct {
-	db      *sql.DB
-	driver  driver
-	dialect string
+	db              *sql.DB
+	driver          driver
+	dialect         string
+	migrationStates map[string]migrationState
 }
 
 // NewManager creates a new schema manager. An error is returned
 // if the provided dialect is unrecognized or if the passed in db is unreachable.
-func NewManager(ctx context.Context, dialect string, db *sql.DB) (*Manager, error) {
+func NewManager(ctx context.Context, dialect string, db *sql.DB, opt ...Option) (*Manager, error) {
 	const op = "schema.NewManager"
 	dbM := Manager{db: db, dialect: dialect}
+	opts := getOpts(opt...)
+	if opts.withMigrationStates != nil {
+		dbM.migrationStates = opts.withMigrationStates
+	} else {
+		// intentionally set it to the reference, so changes to the global var
+		// will be reflected in the manager instance
+		dbM.migrationStates = migrationStates
+	}
 	switch dialect {
 	case "postgres":
 		var err error
@@ -144,7 +154,11 @@ func (b *Manager) RollForward(ctx context.Context) error {
 		return errors.Wrap(err, op)
 	}
 	defer func() {
-		b.driver.Unlock(ctx)
+		if err := b.driver.Unlock(ctx); err != nil {
+			// I'm not sure this is ideal, but we have to rollback the current
+			// transaction if we're unable to release the lock
+			panic(errors.Wrap(err, op))
+		}
 	}()
 
 	curVersion, _, dirty, err := b.driver.CurrentState(ctx)
@@ -156,7 +170,7 @@ func (b *Manager) RollForward(ctx context.Context) error {
 		return errors.New(errors.NotSpecificIntegrity, op, fmt.Sprintf("schema is dirty with version %d", curVersion))
 	}
 
-	if err = b.runMigrations(ctx, newStatementProvider(b.dialect, curVersion)); err != nil {
+	if err = b.runMigrations(ctx, newStatementProvider(b.dialect, curVersion, WithMigrationStates(b.migrationStates))); err != nil {
 		return errors.Wrap(err, op)
 	}
 	return nil
@@ -199,4 +213,49 @@ func (b *Manager) runMigrations(ctx context.Context, qp *statementProvider) erro
 		return errors.Wrap(err, op)
 	}
 	return nil
+}
+
+// LogEntry represents a log entry generated during migrations.
+type LogEntry struct {
+	Id               int
+	MigrationVersion string
+	CreateTime       time.Time
+	Entry            string
+}
+
+// GetMigrationLog will retrieve the migration logs from the db for the last
+// migration. Once it's read the entries, it will delete them from the database.
+//  The WithDeleteLog option is supported and will remove all log entries when provided.
+func GetMigrationLog(ctx context.Context, d *sql.DB, opt ...Option) ([]LogEntry, error) {
+	const op = "schema.GetMigrationLog"
+	const sql = "select id, create_time, migration_version, entry from log_migration where migration_version in (select max(version) from boundary_schema_version)"
+	if d == nil {
+		return nil, errors.New(errors.InvalidParameter, op, "missing sql db")
+	}
+	rows, err := d.QueryContext(ctx, sql)
+	if err != nil {
+		return nil, errors.Wrap(err, op)
+	}
+	defer rows.Close()
+
+	var entries []LogEntry
+	for rows.Next() {
+		var e LogEntry
+		if err := rows.Scan(&e.Id, &e.CreateTime, &e.MigrationVersion, &e.Entry); err != nil {
+			return nil, errors.Wrap(err, op)
+		}
+		entries = append(entries, e)
+	}
+	if rows.Err() != nil {
+		return nil, errors.Wrap(err, op)
+	}
+	opts := getOpts(opt...)
+	if opts.withDeleteLog {
+		// this truncate could change to a delete if FKs are needed in the future
+		_, err = d.ExecContext(ctx, "truncate log_migration")
+		if err != nil {
+			return nil, errors.Wrap(err, op)
+		}
+	}
+	return entries, nil
 }